[move directory] point entity
