+++ src/main/java/kr/hhplus/be/point/entity/Point.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.point;
+package kr.hhplus.be.point.entity;
 
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
