+++ src/main/java/kr/hhplus/be/point/repository/PointHistoryRepository.java
@@ -1,8 +1,6 @@
 package kr.hhplus.be.point.repository;
 
-import kr.hhplus.be.point.Point;
-import kr.hhplus.be.point.PointHistory;
-import kr.hhplus.be.point.TransactionType;
+import kr.hhplus.be.point.entity.PointHistory;
 
 public interface PointHistoryRepository {
     PointHistory save(PointHistory pointHistory);
